+++ waspc/src/Wasp/AppSpec/Api.hs
@@ -48,5 +48,5 @@ method = fst . httpRoute
 path :: Api -> String
 path = snd . httpRoute
 
-data HttpMethod = ALL | GET | POST | PUT | DELETE
+data HttpMethod = ALL | GET | POST | PUT | PATCH | DELETE
   deriving (Show, Eq, Ord, Data, Generic, FromJSON, ToJSON)